product_configurator: rename common test file
